fixed small error had track price as  int not double which was causing no data to be displayed

+++ ITunes Search/Model/Result.swift
@@ -6,7 +6,7 @@ struct Result: Codable {
     let trackName: String
     let releaseDate: String
     let primaryGenreName: String
-    let trackPrice: Int
+    let trackPrice: Double
     
     enum CodingKeys: String, CodingKey {
         case trackId
